@@ -1,1 +1,1 @@
-ruby-2.1.1
+ruby-2.3.0
