@@ -1,2 +1,2 @@
-dummy_migrations_path = File.join(File.dirname(__FILE__), '..', '..', 'db', 'migrate')
-ActiveRecord::Migrator.migrations_path = dummy_migrations_path
+dummy_migrations_paths = File.join(File.dirname(__FILE__), '..', '..', 'db', 'migrate')
+ActiveRecord::Migrator.migrations_paths = dummy_migrations_paths
